UI changes, plust button for testing uploading all products

+++ Stocktakr/src/docketplace/stocktakr/components/QuantityListener.java
@@ -1,5 +1,5 @@
 package docketplace.stocktakr.components;
 
 public interface QuantityListener {
-	public void onChangeQuantity(int newQuantity);
+	public void onChangeQuantity(float newQuantity);
 }
